Remove consum_desc
There's no reason for potions & scrolls to use plus to store
appearance when everything else uses 'appearance'. (Also, there's
no reason for scrolls to use two fields, since it all gets mashed
into one int anyway.)

+++ crawl-ref/source/player.h
@@ -170,7 +170,7 @@ public:
 
   int exp_docked, exp_docked_total; // Ashenzari's wrath
 
-  FixedArray<uint8_t, 6, MAX_SUBTYPES> item_description;
+  FixedArray<uint32_t, 6, MAX_SUBTYPES> item_description;
   FixedVector<unique_item_status_type, MAX_UNRANDARTS> unique_items;
   FixedBitVector<NUM_MONSTERS> unique_creatures;
